Duplicated code moved to sep. module 'items'
