+++ src/aiogram_dialog/widgets/text/list.py
@@ -1,52 +1,23 @@
-from operator import itemgetter
-from typing import Callable, Dict, Sequence, Union
-
-from magic_filter import MagicFilter
+from typing import Dict
 
 from aiogram_dialog.api.protocols import DialogManager
 from aiogram_dialog.widgets.common import WhenCondition
 from .base import Text
-
-ItemsGetter = Callable[[Dict], Sequence]
-
-
-def get_identity(items: Sequence) -> ItemsGetter:
-    def identity(data) -> Sequence:
-        return items
-
-    return identity
-
-
-def get_magic_getter(f: MagicFilter) -> ItemsGetter:
-    def items_magic(data: Dict) -> Sequence:
-        items = f.resolve(data)
-        if isinstance(items, Sequence):
-            return items
-        else:
-            return []
-
-    return items_magic
+from ..common.items import ItemsGetterVariant, get_items_getter
 
 
 class List(Text):
     def __init__(
             self,
             field: Text,
-            items: Union[str, ItemsGetter, MagicFilter, Sequence],
+            items: ItemsGetterVariant,
             sep: str = "\n",
             when: WhenCondition = None,
     ):
         super().__init__(when=when)
         self.field = field
         self.sep = sep
-        if isinstance(items, str):
-            self.items_getter = itemgetter(items)
-        elif isinstance(items, MagicFilter):
-            self.items_getter = get_magic_getter(items)
-        elif isinstance(items, Callable):
-            self.items_getter = items
-        else:
-            self.items_getter = get_identity(items)
+        self.items_getter = get_items_getter(items)
 
     async def _render_text(
             self, data: Dict, manager: DialogManager,
